backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "asyncpg>=0.30.0",
    "bcrypt>=4.3.0",
    "fastapi[standard]>=0.115.12",
    "orjson>=3.10.16",
    "passlib>=1.7.4",
    "pydantic-settings>=2.9.1",
    "pyjwt>=2.10.1",
    "python-dotenv>=1.1.0",
    "sqlalchemy>=2.0.40",
]

[dependency-groups]
dev = [
    "isort>=6.0.1",
    "ruff>=0.11.6",
    "taskipy>=1.14.1",
    "types-passlib>=1.7.7.20250408",
]

[tool.taskipy.tasks]
dev = "uv run main.py"
lint = "ruff check ."
format = "ruff format ."
test = "pytest"
sort = "isort ."
